add image func

+++ app/Http/Controllers/ImageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Image;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class ImageController extends Controller
 {
@@ -32,7 +33,12 @@ class ImageController extends Controller
     }
 
     function destroy(Image $image) {
+        if (!$image->path)
+            return redirect()->route('product.index');
+
         $image->delete();
+        Storage::delete($image->path);
+
         return redirect()->route('product.index');
     }
 }
